Renaming the interface files
